attendance-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
dlib==19.24.6
face-recognition==1.3.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
trimesh>=4.4.0
